security: restrict Discord infrastructure commands
